Cleaned the logic on some of the hashtable code.
git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/X86ElfInstrumentor@1898 55925d41-9515-0410-9586-c6f35bcc715f

+++ include/Base.h
@@ -121,7 +121,7 @@ typedef void (*fprintf_ftype)(FILE*, const char*, ...);
 #define Size__32_bit_Dynamic_Entry          sizeof(Elf32_Dyn)
 #define Size__64_bit_Dynamic_Entry          sizeof(Elf64_Dyn)
 #define Size__32_bit_Hash_Entry             sizeof(uint32_t)
-#define Size__64_bit_Hash_Entry             sizeof(uint64_t)
+#define Size__64_bit_Hash_Entry             sizeof(uint32_t)
 
 #define Size__32_bit_ExceptionTable_Entry     EXCEPTSZ
 #define Size__64_bit_ExceptionTable_Entry     EXCEPTSZ_64
